Safari-609.1.11.1/LayoutTests/media/controls/airplay-controls-expected.txt

This tests that showing the wireless device placeholder does not show controls.

This test only runs in DRT!


EVENT: canplaythrough

Test initial state

PASS: We are using the apple idiom
PASS: video.controls should be false
PASS: Controls should not be visible

Select an AppleTV device

PASS: Placeholder is visible
PASS: Placeholder has classname 'appletv'
PASS: Controls should not be visible

Testing finished.
